Move most of binwalk's tinfl additions to a separate file
This will make downstream use of unbundled tinfl easier.
